feat: 검증된 BlackJack 클래스 기능 구현
