+++ README.md
@@ -25,10 +25,10 @@
     - [x] 손패는 카드들의 숫자 총합을 반환한다.
 - 블랙잭
   - 블랙잭 모델은 손패와 참여자의 상태를 나타내는 state를 가지고 있다.
-    - [ ] 블랙잭에서는 손패의 점수를 판단하여 state를 변환한다.
-      - [ ] 손패의 합이 21이 넘으면 `버스트` 상태로 변환한다.
-      - [ ] 플레이어는 합이 21 미만이면 `히트` 또는 `스테이` 상태를 결정할 수 있다.
-      - [ ] 플레이어는 손패의 합이 21이라면 블랙잭 상태가 된다.
+    - [x] 블랙잭에서는 손패의 점수를 판단하여 state를 변환한다.
+      - [x] 손패의 합이 21이 넘으면 `버스트` 상태로 변환한다.
+      - [x] 플레이어는 합이 21 미만이면 `히트` 또는 `스테이` 상태를 결정할 수 있다.
+      - [x] 플레이어는 손패의 합이 21이라면 블랙잭 상태가 된다.
 ---
 - State
   - state는 참여자의 진행 상태를 나타낸다.
