--- src/main/kotlin/blackjack/state/Action.kt
@@ -1,5 +1,0 @@
-package blackjack.state
-
-sealed interface Action : State {
-    data object Hit : Action
-}
