--- src/main/kotlin/blackjack/state/Finish.kt
@@ -1,7 +1,0 @@
-package blackjack.state
-
-sealed interface Finish : State {
-    data object Bust : Finish
-    data object Stay : Finish
-    data object BlackJack : Finish
-}
