+++ src/main/kotlin/blackjack/state/State.kt
@@ -1,3 +1,15 @@
 package blackjack.state
 
-sealed interface State
+sealed interface State {
+
+    sealed interface Action : State {
+        data object Hit : Action
+    }
+
+    sealed interface Finish : State {
+        data object Bust : Finish
+        data object Stay : Finish
+        data object BlackJack : Finish
+    }
+
+}
